requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-jsonapi>=6.1.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
